requirements.txt
pytesseract
pillow 
google-genai 
opencv-python
python-dotenv
streamlit
groq